# Optimize std_hash in coin.py