--- a/chia/util/hash.py
+++ b/chia/util/hash.py
@@ -1,10 +1,13 @@
-import blspy
+from hashlib import sha256
 
 from chia.types.blockchain_format.sized_bytes import bytes32
 
 
-def std_hash(b) -> bytes32:
+def std_hash(b, skip_bytes_conversion: bool = False) -> bytes32:
     """
     The standard hash used in many places.
     """
-    return bytes32(blspy.Util.hash256(bytes(b)))
+    if skip_bytes_conversion:
+        return bytes32(sha256(b).digest())
+    else:
+        return bytes32(sha256(bytes(b)).digest())