constructing a bytes object from a bytes object is not necessary. In these cases we are passing in bytes so we don't need to convert. These are also the places where `std_hash` is called the most.

Also, using hashlib instead of calling into C++ is more efficient.

This makes `std_hash` around 25% faster

test_full_sync
| before | after | after/before |
| --- | --- | --- |
| 119s | 116s |  0.975 |

Another thing to improve would be not checking the size when constructing a bytes32 here. It's very slow for some reason (calling len(bytes)). However I'm not sure how to add an optional arg to the constructor of bytes32