@arvidn 2022-05-17
I'm experimenting with a more radincal performance improvement to coin. I'm tempted to suggest holding this off a bit. It's a pretty small speed-up too (maybe even within the error margins).

I suspect `std_hash` releases the GIL, and that's why it's so much slower than the python-native hash function.