@mariano54 2022-05-17
I ran multiple times so I'm confident in the speedup. Also seems like a harmless change.  If this is addressed in your other changes, we can close