@mariano54 2022-05-24
@arvidn do you plan on making those Coin changes? If so, let's close. 
If not, I think we should merge this one and you can build on top when you come back to it