@arvidn 2022-05-24
yeah, it probably makes sense to merge this anyway. It's a step in the right direction. My coin patch depends on https://github.com/Chia-Network/chia-blockchain/pull/11503 so it will probably be at least a week before it can land.